requirements.txt
bidict==0.22.1
blinker==1.6.2
click==8.1.7
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==0.28.1
langchain==0.0.330
faiss-cpu==1.7.4
pytz
Werkzeug==2.3.7
pypdf2
anthropic
langdetect
python-dotenv
tiktoken